playwright/requirements.txt
# Core dependencies
playwright==1.40.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
openpyxl==3.11.0

# Optional dependencies
pytest-cov==4.1.0
